+++ packages/diagram-protocol/src/diagram/transaction/defaultEditTypes.ts
